+++ src/events/ImageTransformerOperationEvent.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ * @link https://craftcms.com/
+ * @copyright Copyright (c) Pixel & Tonic, Inc.
+ * @license https://craftcms.github.io/license/
+ */
+
+namespace craft\events;
+
+use craft\elements\Asset;
+use craft\models\ImageTransformIndex;
+use yii\base\Event;
+
+/**
+ * Image transformer operation event class.
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 4.0.0
+ */
+class ImageTransformerOperationEvent extends Event
+{
+    /**
+     * @var Asset The Asset that is associated with the event
+     */
+    public Asset $asset;
+
+    /**
+     * @var ImageTransformIndex The image transform model associated with the event.
+     */
+    public ImageTransformIndex $imageTransformIndex;
+
+    /**
+     * @var string The resulting file path that will either be created or deleted.
+     */
+    public string $path;
+}